Restructure CPU
The entire CPU layout was redone to reflect a more realistic, final
layout. The opcodes were also created and added to the CPU so that
tests could be performed. Reading from the instruction cache and
Registers has been setup and tested, writing has not. Many circuits
were redone internally to reflect the changes in design. For example,
the ALU was given a NOP command to pass the ALU-R to ALU-O this is used
in when either executing STORE or LOAD. Additionally the bytecode has
to be the mirror image of the assembly ie:

ADD R0 R1 = 0001 0 110 instead of 011 0 1000

The CPU is still clockless since I haven't created the buffer registers
necessary to make everything move in lockstep. At this point it is a
lot closer to a full pipeline CPU than a "pipeline" classical CPU.